Improve loading of sigma in epf.
About 2.5% speed increase when doing EPF0+EPF1.

Also split up EPF stages in multiple files.
